move back to table.tsx
